requirements.txt
# ── Core ML ──────────────────────────────────────────────
lightgbm>=4.3.0
xgboost>=2.0.3
catboost>=1.2.5
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0

# ── Optimisation ─────────────────────────────────────────
optuna>=3.6.0
optuna-integration[lightgbm]>=3.6.0

# ── MLflow ───────────────────────────────────────────────
mlflow>=2.12.0

# ── Drift detection ──────────────────────────────────────
evidently>=0.4.30

# ── Interpretability ─────────────────────────────────────
shap>=0.45.0

# ── API ──────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
python-multipart>=0.0.9

# ── Frontend ─────────────────────────────────────────────
streamlit>=1.32.0
plotly>=5.20.0
requests>=2.31.0

# ── Serialization ────────────────────────────────────────
joblib>=1.3.2

# ── Tests ────────────────────────────────────────────────
pytest>=8.1.0
httpx>=0.27.0
pytest-asyncio>=0.23.0

# ── Utils ────────────────────────────────────────────────
python-dotenv>=1.0.0
pyyaml>=6.0.1
loguru>=0.7.2